+++ sourceCode/geometricAlgorithms/edgeElement.h
@@ -6,7 +6,7 @@
 class CEdgeElement  
 {
 public:
-    CEdgeElement(int v0,int v1,int tri,CEdgeElement* nextEl,int position,float vertices[]);
+    CEdgeElement(int v0,int v1,int tri,CEdgeElement* nextEl,int position,double vertices[]);
     virtual ~CEdgeElement();
     int vertex0;
     int vertex1;
