+++ sourceCode/geometricAlgorithms/meshRoutines.h
@@ -7,30 +7,30 @@
 class CMeshRoutines  
 {
 public:
-    static int convexDecompose(const float* vertices,int verticesLength,const int* indices,
-                               int indicesLength,std::vector<std::vector<float>*>& verticesList,
+    static int convexDecompose(const double* vertices,int verticesLength,const int* indices,
+                               int indicesLength,std::vector<std::vector<double>*>& verticesList,
                                std::vector<std::vector<int>*>& indicesList,size_t nClusters,
                                double concavity,bool addExtraDistPoints,bool addFacesPoints,
                                double ccConnectDist,size_t targetNTrianglesDecimatedMesh,
                                size_t maxHullVertices,double smallestClusterThreshold,
-                               bool useHACD,int resolution_VHACD,int depth_VHACD_old,float concavity_VHACD,
+                               bool useHACD,int resolution_VHACD,int depth_VHACD_old,double concavity_VHACD,
                                int planeDownsampling_VHACD,int convexHullDownsampling_VHACD,
-                               float alpha_VHACD,float beta_VHACD,float gamma_VHACD_old,bool pca_VHACD,
-                               bool voxelBased_VHACD,int maxVerticesPerCH_VHACD,float minVolumePerCH_VHACD);
+                               double alpha_VHACD,double beta_VHACD,double gamma_VHACD_old,bool pca_VHACD,
+                               bool voxelBased_VHACD,int maxVerticesPerCH_VHACD,double minVolumePerCH_VHACD);
 
-    static void getEdgeFeatures(float* vertices,int verticesLength,int* indices,int indicesLength,
-            std::vector<int>* vertexIDs,std::vector<int>* edgeIDs,std::vector<int>* faceIDs,float angleTolerance,bool forDisplay,bool hideEdgeBorders);
-    static bool getConvexHull(std::vector<float>* verticesInOut,std::vector<int>* indicesInOut);
-    static bool getConvexHull(const std::vector<float>* verticesIn,std::vector<float>* verticesOut,std::vector<int>* indicesOut);
-    static bool getConvexHull(const float* verticesIn,int verticesInLength,std::vector<float>* verticesOut,std::vector<int>* indicesOut);
-    static bool getDecimatedMesh(const std::vector<float>& verticesIn,const std::vector<int>& indicesIn,float decimationPercentage,std::vector<float>& verticesOut,std::vector<int>& indicesOut);
+    static void getEdgeFeatures(double* vertices,int verticesLength,int* indices,int indicesLength,
+            std::vector<int>* vertexIDs,std::vector<int>* edgeIDs,std::vector<int>* faceIDs,double angleTolerance,bool forDisplay,bool hideEdgeBorders);
+    static bool getConvexHull(std::vector<double>* verticesInOut,std::vector<int>* indicesInOut);
+    static bool getConvexHull(const std::vector<double>* verticesIn,std::vector<double>* verticesOut,std::vector<int>* indicesOut);
+    static bool getConvexHull(const double* verticesIn,int verticesInLength,std::vector<double>* verticesOut,std::vector<int>* indicesOut);
+    static bool getDecimatedMesh(const std::vector<double>& verticesIn,const std::vector<int>& indicesIn,double decimationPercentage,std::vector<double>& verticesOut,std::vector<int>& indicesOut);
 
     inline static bool getMinDistBetweenSegmentAndPoint_IfSmaller(const C3Vector& lp0,
-                            const C3Vector& lv0,const C3Vector& dummyPos,float &dist,C3Vector& segA)
+                            const C3Vector& lv0,const C3Vector& dummyPos,double &dist,C3Vector& segA)
     {   // dist & segA are modified only if the distance is smaller than 'dist' which was input
         // Return value of true means: the dist is smaller
         // The segment can be degenerated (point)
-        float d;
+        double d;
         if ((lv0(0)==0.0)&&(lv0(1)==0.0)&&(lv0(2)==0.0))
         { // We have a degenerated segment here (point):
             C3Vector v(dummyPos-lp0);
@@ -44,7 +44,7 @@ public:
             return(false);
         }
         // dist between lines described by segment and dummyPos
-        float t=getMinDistPtBetweenPointAndLine(dummyPos,lp0,lv0);
+        double t=getMinDistPtBetweenPointAndLine(dummyPos,lp0,lv0);
         if ((t>=0.0)&&(t<=1.0))
         {
             C3Vector v(dummyPos-lp0-(lv0*t));
@@ -78,22 +78,22 @@ public:
         return(smaller);
     }
 
-    inline static float getMinDistPtBetweenPointAndLine(const C3Vector& point,const C3Vector& lp,const C3Vector& lv)
+    inline static double getMinDistPtBetweenPointAndLine(const C3Vector& point,const C3Vector& lp,const C3Vector& lv)
     {
         C3Vector d(point-lp);
         return((d*lv)/(lv*lv));
     };
 
-    static bool checkIfConvex(const std::vector<float>& vertices,const std::vector<int>& indices,float distanceToleranceInPercent);
+    static bool checkIfConvex(const std::vector<double>& vertices,const std::vector<int>& indices,double distanceToleranceInPercent);
 
-    static void createCube(std::vector<float>& vertices,std::vector<int>& indices,const C3Vector& sizes,const int subdivisions[3]);
-    static void createSphere(std::vector<float>& vertices,std::vector<int>& indices,const C3Vector& sizes,int sides,int faces);
-    static void createCapsule(std::vector<float>& vertices,std::vector<int>& indices,const C3Vector& sizes,int sides,int faceSubdiv);
-    static void createCylinder(std::vector<float>& vertices,std::vector<int>& indices,const C3Vector& sizes,int sides,int faces,int discDiv,bool openEnds,bool cone);
-    static void createAnnulus(std::vector<float>& vertices,std::vector<int>& indices,float Dlarge,float Dsmall,float zShift,int sides,bool faceUp);
+    static void createCube(std::vector<double>& vertices,std::vector<int>& indices,const C3Vector& sizes,const int subdivisions[3]);
+    static void createSphere(std::vector<double>& vertices,std::vector<int>& indices,const C3Vector& sizes,int sides,int faces);
+    static void createCapsule(std::vector<double>& vertices,std::vector<int>& indices,const C3Vector& sizes,int sides,int faceSubdiv);
+    static void createCylinder(std::vector<double>& vertices,std::vector<int>& indices,const C3Vector& sizes,int sides,int faces,int discDiv,bool openEnds,bool cone);
+    static void createAnnulus(std::vector<double>& vertices,std::vector<int>& indices,double Dlarge,double Dsmall,double zShift,int sides,bool faceUp);
 
 private:
     static void _insertEdge(std::vector<std::vector<int>* >& allEdges,int vertexIndex1,int vertexIndex2,int triangleIndex); // used for convex check
     static int _getTriangleIndexFromEdge(std::vector<std::vector<int>* >& allEdges,int vertexIndex1,int vertexIndex2,int triangleIndexToExclude); // used for convex check
-    static bool _removeColinearTrianglePoints(std::vector<float>* vertices,std::vector<int>* indices,float toleranceDist);
+    static bool _removeColinearTrianglePoints(std::vector<double>* vertices,std::vector<int>* indices,double toleranceDist);
 };
